[refactor] formTextField 라는 최상위 클래스 만들어서 서브클래스들이 모두 초기 세팅이 	  적용 되도록 기능 구현함 [feat] 로그인관련 텍스트 필드 만듬
