+++ iOS/SignupApp/SignupApp/Views/TextFields/LoginTextFields/LoginTextField.swift
@@ -1,0 +1,13 @@
+//
+//  LoginTextField.swift
+//  SignupApp
+//
+//  Created by kimdo2297 on 2020/03/28.
+//  Copyright © 2020 Jason. All rights reserved.
+//
+
+import UIKit
+
+class LoginTextField: FormTextField {
+    
+}
